requirements.txt
Django==4.1.1
python-decouple==3.6
psycopg2-binary>=2.8
psycopg2==2.8.6
djangorestframework==3.13.1
drf-yasg==1.21.3
drfdocs==0.0.11
swagger-spec-validator==2.7.6
django-filter==22.1
traceback-with-variables==2.0.4
faker==15.1.1
pytest==7.1.3
pytest-django==4.5.2